+++ tests/testthat/test-do_RidgePlot.R
@@ -13,27 +13,8 @@ if(base::isFALSE(dep_check[["do_RidgePlot"]])){
                               viridis.direction = 1)
     testthat::expect_type(p, "list")
 
-    p <- SCpubr::do_RidgePlot(sample = sample,
-                              feature = "nFeature_RNA",
-                              continuous_scale = TRUE,
-                              compute_quantiles = TRUE,
-                              compute_distribution_tails = TRUE)
-    testthat::expect_type(p, "list")
 
-    p <- SCpubr::do_RidgePlot(sample = sample,
-                              feature = "nCount_RNA",
-                              continuous_scale = TRUE,
-                              compute_quantiles = TRUE,
-                              compute_distribution_tails = TRUE,
-                              prob_tails = 0.4)
-    testthat::expect_type(p, "list")
 
-    p <- SCpubr::do_RidgePlot(sample = sample,
-                              feature = "nFeature_RNA",
-                              continuous_scale = TRUE,
-                              compute_quantiles = TRUE,
-                              color_by_probabilities = TRUE)
-    testthat::expect_type(p, "list")
 
   })
 
@@ -68,42 +49,8 @@ if(base::isFALSE(dep_check[["do_RidgePlot"]])){
                               sequential.direction = -1)
     testthat::expect_type(p, "list")
     
-    p <- SCpubr::do_RidgePlot(sample = sample,
-                              feature = "nFeature_RNA",
-                              continuous_scale = TRUE,
-                              compute_quantiles = TRUE,
-                              color_by_probabilities = TRUE,
-                              use_viridis = TRUE,
-                              viridis.direction = 1)
-    testthat::expect_type(p, "list")
-    
-    p <- SCpubr::do_RidgePlot(sample = sample,
-                              feature = "nFeature_RNA",
-                              continuous_scale = TRUE,
-                              compute_quantiles = TRUE,
-                              color_by_probabilities = TRUE,
-                              use_viridis = TRUE,
-                              viridis.direction = -1)
-    testthat::expect_type(p, "list")
-    
-    p <- SCpubr::do_RidgePlot(sample = sample,
-                              feature = "nFeature_RNA",
-                              continuous_scale = TRUE,
-                              compute_quantiles = TRUE,
-                              color_by_probabilities = TRUE,
-                              use_viridis = FALSE,
-                              sequential.direction = 1)
-    testthat::expect_type(p, "list")
-    
-    p <- SCpubr::do_RidgePlot(sample = sample,
-                              feature = "nFeature_RNA",
-                              continuous_scale = TRUE,
-                              compute_quantiles = TRUE,
-                              color_by_probabilities = TRUE,
-                              use_viridis = FALSE,
-                              sequential.direction = -1)
-    testthat::expect_type(p, "list")
-    
+
+
     p <- SCpubr::do_RidgePlot(sample = sample,
                               feature = "nCount_RNA")
     testthat::expect_type(p, "list")
@@ -186,57 +133,8 @@ if(base::isFALSE(dep_check[["do_RidgePlot"]])){
     testthat::expect_type(p, "list")
   })
 
-  testthat::test_that("do_RidgePlot: PASS - quantiles", {
-    testthat::skip_on_cran()
 
 
-    p <- SCpubr::do_RidgePlot(sample = sample,
-                              feature = "nCount_RNA",
-                              continuous_scale = TRUE,
-                              compute_custom_quantiles = TRUE,
-                              compute_quantiles = TRUE)
-    testthat::expect_type(p, "list")
-
-    p <- SCpubr::do_RidgePlot(sample = sample,
-                              feature = "nCount_RNA",
-                              continuous_scale = TRUE,
-                              compute_custom_quantiles = TRUE,
-                              compute_quantiles = TRUE,
-                              quantiles = c(0.1, 0.5, 0.9))
-    testthat::expect_type(p, "list")
-  })
-
-  testthat::test_that("do_RidgePlot: PASS - distribution tails", {
-    testthat::skip_on_cran()
-
-
-    p <- SCpubr::do_RidgePlot(sample = sample,
-                              feature = "nFeature_RNA",
-                              continuous_scale = TRUE,
-                              compute_quantiles = TRUE,
-                              compute_distribution_tails = TRUE)
-    testthat::expect_type(p, "list")
-
-    p <- SCpubr::do_RidgePlot(sample = sample,
-                              feature = "nCount_RNA",
-                              continuous_scale = TRUE,
-                              compute_quantiles = TRUE,
-                              compute_distribution_tails = TRUE,
-                              prob_tails = 0.4)
-    testthat::expect_type(p, "list")
-  })
-
-  testthat::test_that("do_RidgePlot: PASS - distribution tails", {
-    testthat::skip_on_cran()
-
-
-    p <- SCpubr::do_RidgePlot(sample = sample,
-                              feature = "nFeature_RNA",
-                              continuous_scale = TRUE,
-                              compute_quantiles = TRUE,
-                              color_by_probabilities = TRUE)
-    testthat::expect_type(p, "list")
-  })
 
   testthat::test_that("do_RidgePlot: PASS - flip", {
     testthat::skip_on_cran()